+++ src/assets/migrations/create-table.js
@@ -1,7 +1,7 @@
 'use strict';
 
 module.exports = {
-  up: function (queryInterface, Sequelize) {
+  up: (queryInterface, Sequelize) => {
     return queryInterface
       .createTable('<%= tableName %>', {
         id: {
@@ -29,7 +29,7 @@ module.exports = {
       });
   },
 
-  down: function (queryInterface, Sequelize) {
+  down: (queryInterface, Sequelize) => {
     return queryInterface.dropTable('<%= tableName %>');
   }
 };