+++ src/assets/models/model.js
@@ -1,6 +1,6 @@
 'use strict';
 
-module.exports = function(sequelize, DataTypes) {
+module.exports = (sequelize, DataTypes) => {
   var <%= name %> = sequelize.define('<%= name %>', {
     <% Object.keys(attributes).forEach(function(fieldName, index) { %>
       <%= fieldName %>: DataTypes.<%= attributes[fieldName].toUpperCase() %>