+++ test/seed/create.test.js
@@ -48,8 +48,8 @@ const gulp      = require('gulp');
           .src(Support.resolveSupportPath('tmp', 'seeders'))
           .pipe(helpers.readFile('*-' + seedFile))
           .pipe(helpers.expect(stdout => {
-            expect(stdout).to.contain('up: function (queryInterface, Sequelize) {');
-            expect(stdout).to.contain('down: function (queryInterface, Sequelize) {');
+            expect(stdout).to.contain('up: (queryInterface, Sequelize) => {');
+            expect(stdout).to.contain('down: (queryInterface, Sequelize) => {');
           }))
           .pipe(helpers.teardown(done));
       });